requirements.txt
notion_client==2.2.1
python-dotenv==1.0.0
pydantic==2.5.3
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
